Add sta_list.sh to filter SAC files for stacking
Use -h or --help options for available filters by which to
limit the stations included in the output list.

By default, file names are quoted.  This is so that if they
contain path separators ('/'s), this does not upset the
Fortran programs which do the stacking.  A / on input will
be read as needing to start a new line, which isn't desired.