Add filtering options to stack_fk
Build the code with `make filt` to link to your system's
libxapiir, which is bundled with SAC.  (You may need to set
$LIB_XAPIIR with the correct linker options.)  This enables
one to use the -bp, -hp and -lp options to filter the SAC
data before stacking within the program.

It is recommended that any tapering be pre-applied to
avoid artifacts near the beginning or end of time series.