Add filtering to stack_vespa
When built against the SAC libxapiir library, one can
pass -bp, -hp or -lp options to perform filtering on
the SAC traces before stacking.  Build with
`make filt` to enable this, if necessary adjusting
the $LIB_XAPIIR variable with the correct linker
options.

It is advised to perform tapering and other adjustments
to the traces beforehand to avoid artifacts.